Associate clients in pairs as they connect
Uses socket.io (http://socket.io/) to send messages between clients and
server.

When one client connects, the server remembers it, waits for a second
one to connect, then pairs them. The clients are notified when they are
matched and when their corresponding client disconnects. Clients are
associated by assigning their corresponding socket to a room, which is
named using a counter.